+++ samples/trade360-samples/src/main/java/eu/lsports/trade360feedexample/handlers/inplay/messages/HeartbeatHandlerInplay.java
@@ -11,7 +11,8 @@ public class HeartbeatHandlerInplay implements EntityHandler<HeartbeatUpdate> {
     protected static final Logger logger = LogManager.getLogger();
     @Override
     public void process(HeartbeatUpdate entity, Map<String, String> messageHeaders, Map<String, String> transportMessageHeaders) {
-        logger.info("Received: " + this.getClass().getSimpleName());
+        logger.info("[FIH] HeartbeatUpdate InPlay: feedInterruptedDomains={} (empty=normal)",
+                entity == null ? "[]" : java.util.Arrays.toString(entity.feedInterrupted));
     }
 
     @Override